requirements.txt
pandas
numpy
matplotlib
seaborn
ipykernel
pathlib
langchain-core
langchain-community
langchain-text-splitters
sentence-transformers
faiss-cpu
scikit-learn
uvicorn
fastapi